Refactor: repository interface を domain 層に移動
DDD の依存性逆転の原則に従い、repository interface を
infrastructure 層から domain 層に移動

- shared/domain/repositories/ に共通リポジトリ interface を配置
- report/domain/repositories/ にレポート用リポジトリ interface を配置
- infrastructure 層が domain 層に依存する正しい依存関係に修正
