+++ admin/src/server/contexts/data-import/application/usecases/save-preview-transactions-usecase.ts
@@ -4,7 +4,7 @@ import type {
   CreateTransactionInput,
   UpdateTransactionInput,
 } from "@/server/contexts/shared/domain/transaction";
-import type { ITransactionRepository } from "@/server/contexts/shared/infrastructure/repositories/interfaces/transaction-repository.interface";
+import type { ITransactionRepository } from "@/server/contexts/shared/domain/repositories/transaction-repository.interface";
 import type { PreviewTransaction } from "@/server/contexts/data-import/domain/models/preview-transaction";
 import { MfRecordConverter } from "@/server/contexts/data-import/infrastructure/mf/mf-record-converter";
