+++ admin/src/server/contexts/report/application/services/income-assembler.ts
@@ -6,7 +6,7 @@
  */
 import "server-only";
 
-import type { IReportTransactionRepository } from "@/server/contexts/report/infrastructure/repositories/interfaces/report-transaction-repository.interface";
+import type { IReportTransactionRepository } from "@/server/contexts/report/domain/repositories/report-transaction-repository.interface";
 import {
   convertToBusinessIncomeSection,
   convertToGrantIncomeSection,
