+++ admin/src/server/contexts/shared/application/usecases/create-balance-snapshot-usecase.ts
@@ -2,7 +2,7 @@ import type {
   BalanceSnapshot,
   CreateBalanceSnapshotInput,
 } from "@/shared/models/balance-snapshot";
-import type { IBalanceSnapshotRepository } from "@/server/contexts/shared/infrastructure/repositories/interfaces/balance-snapshot-repository.interface";
+import type { IBalanceSnapshotRepository } from "@/server/contexts/shared/domain/repositories/balance-snapshot-repository.interface";
 
 export class CreateBalanceSnapshotUsecase {
   constructor(private repository: IBalanceSnapshotRepository) {}
