+++ admin/src/server/contexts/shared/domain/repositories/political-organization-repository.interface.ts
