+++ admin/src/server/contexts/shared/domain/repositories/user-repository.interface.ts
