+++ admin/tests/server/contexts/data-import/application/usecases/save-preview-transactions-usecase.test.ts
@@ -8,7 +8,7 @@ import {
   PreviewMfCsvUsecase,
   type PreviewMfCsvInput,
 } from "@/server/contexts/data-import/application/usecases/preview-mf-csv-usecase";
-import type { ITransactionRepository } from "@/server/contexts/shared/infrastructure/repositories/interfaces/transaction-repository.interface";
+import type { ITransactionRepository } from "@/server/contexts/shared/domain/repositories/transaction-repository.interface";
 import type { CreateTransactionInput } from "@/server/contexts/shared/domain/transaction";
 
 describe("SavePreviewTransactionsUsecase", () => {
