+++ admin/tests/server/contexts/report/application/services/income-assembler.test.ts
@@ -2,7 +2,7 @@ import {
   IncomeAssembler,
   type IncomeAssemblerInput,
 } from "@/server/contexts/report/application/services/income-assembler";
-import type { IReportTransactionRepository } from "@/server/contexts/report/infrastructure/repositories/interfaces/report-transaction-repository.interface";
+import type { IReportTransactionRepository } from "@/server/contexts/report/domain/repositories/report-transaction-repository.interface";
 import type {
   BusinessIncomeTransaction,
   GrantIncomeTransaction,
